Add facade layer, change DTOs, implement dish related APIs, initialize ingredient related services, implement ingredient create/add/remove APIs

+++ wte-backend/src/main/java/net/summer23project/wtebackend/entity/DishIngredientAmount.java
@@ -24,7 +24,7 @@ public class DishIngredientAmount {
     private Long id;
 
     @Column(name = "ingredient_amount", nullable = false)
-    private int ingredientAmount;
+    private double ingredientAmount;
 
     @ManyToOne
     @JoinColumn(name = "dish_id")
